Dot Placement and Ghost Upgrades
Implemented proper placement of all dots on the board as well as added in the remaining ghosts with their own classes to implement their unique behaviors. Also added the graphics for the spawn area and the colors for the ghosts.  As well as the spawning feature of the ghosts in proper order.